sharding-and-partitioning/requirements.txt
mysql-connector-python==8.0.26
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==1.0.1